@@ -122,47 +113,47 @@
     "@commitlint/config-conventional": "^19.8.1",
     "@eslint/eslintrc": "^3.2.0",
     "@eslint/js": "^9.20.0",
-    "@nestjs/cli": "11.0.7",
-    "@nestjs/schematics": "11.0.5",
-    "@nestjs/testing": "11.1.2",
+    "@nestjs/cli": "11.0.10",
+    "@nestjs/schematics": "11.0.9",
+    "@nestjs/testing": "11.1.6",
     "@release-it/conventional-changelog": "10.0.1",
-    "@swc/cli": "0.6.0",
-    "@swc/core": "1.11.21",
-    "@types/express": "5.0.1",
-    "@types/jest": "29.5.14",
-    "@types/minio": "^7.1.0",
+    "@swc/cli": "0.7.8",
+    "@swc/core": "1.13.4",
+    "@types/bcryptjs": "3.0.0",
+    "@types/express": "5.0.3",
+    "@types/jest": "30.0.0",
     "@types/ms": "2.1.0",
-    "@types/multer": "1.4.12",
-    "@types/node": "22.15.21",
+    "@types/multer": "2.0.0",
+    "@types/node": "22.18.3",
     "@types/passport-anonymous": "1.0.5",
     "@types/passport-jwt": "4.0.1",
     "@types/socket.io": "^3.0.1",
     "@types/socket.io-client": "^1.4.36",
     "@types/supertest": "6.0.3",
     "@types/ws": "^8.18.1",
-    "@typescript-eslint/eslint-plugin": "8.29.1",
-    "@typescript-eslint/parser": "8.29.1",
+    "@typescript-eslint/eslint-plugin": "8.44.0",
+    "@typescript-eslint/parser": "8.44.0",
     "env-cmd": "10.1.0",
-    "eslint": "9.27.0",
-    "eslint-config-prettier": "10.1.5",
-    "eslint-plugin-prettier": "5.2.6",
+    "eslint": "9.35.0",
+    "eslint-config-prettier": "10.1.8",
+    "eslint-plugin-prettier": "5.5.4",
     "globals": "^16.0.0",
     "husky": "^9.1.7",
     "hygen": "6.2.11",
     "is-ci": "4.1.0",
-    "jest": "29.7.0",
+    "jest": "30.1.3",
     "lint-staged": "^16.2.6",
     "ntl": "^5.1.0",
-    "prettier": "3.5.3",
+    "prettier": "3.6.2",
     "prompts": "2.4.2",
-    "release-it": "19.0.2",
-    "supertest": "7.1.1",
-    "ts-jest": "29.3.4",
-    "ts-loader": "9.5.2",
+    "release-it": "19.0.5",
+    "supertest": "7.1.4",
+    "ts-jest": "29.4.1",
+    "ts-loader": "9.5.4",
     "ts-node": "10.9.2",
     "tsconfig-paths": "4.2.0",
     "tslib": "2.8.1",
-    "typescript": "5.8.3"
+    "typescript": "5.9.2"
   },
   "jest": {
     "moduleFileExtensions": [
